+++ example/ios/Flutter/flutter_export_environment.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
 export "FLUTTER_ROOT=/Users/matin/flutter-development/flutter"
-export "FLUTTER_APPLICATION_PATH=/Users/matin/Documents/fitgo/flutter-plugins/packages/pedometer/example"
+export "FLUTTER_APPLICATION_PATH=/Users/matin/Documents/fitgo/daily_pedometer2/example"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_TARGET=lib/main.dart"
 export "FLUTTER_BUILD_DIR=build"