+++ example/ios/Runner/GeneratedPluginRegistrant.m
@@ -6,8 +6,8 @@
 
 #import "GeneratedPluginRegistrant.h"
 
-#if __has_include(<daily_pedometer2/PedometerPlugin.h>)
-#import <daily_pedometer2/PedometerPlugin.h>
+#if __has_include(<daily_pedometer2/DailyPedometer2Plugin.h>)
+#import <daily_pedometer2/DailyPedometer2Plugin.h>
 #else
 @import daily_pedometer2;
 #endif
@@ -15,7 +15,7 @@
 @implementation GeneratedPluginRegistrant
 
 + (void)registerWithRegistry:(NSObject<FlutterPluginRegistry>*)registry {
-  [PedometerPlugin registerWithRegistrar:[registry registrarForPlugin:@"PedometerPlugin"]];
+  [DailyPedometer2Plugin registerWithRegistrar:[registry registrarForPlugin:@"DailyPedometer2Plugin"]];
 }
 
 @end