+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.7
+
+-   ios & android plugins updated
+-   ios & android plugins bug fix
+
 ## 1.0.6
 
 -   Forked The Pedometer Plugin